@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   extends: 'lighthouse:default',
   settings: {
     onlyAudits: ['first-meaningful-paint', 'speed-index', 'interactive']
